seperated combined header file implementations into seperate .cpp and .hpp files
